refactor(refactoring): refactoring after codereview
I removed the extra code after introducing the context, corrected the comments after the code review
