+++ src/App.jsx
@@ -2,7 +2,7 @@ import React from "react";
 import { HomeWithAuth } from "./pages/Home";
 import Welcome from "./pages/Welcome";
 import "./App.css";
-import { withAuth } from "./components/context/withAuth";
+import { withAuth } from "./context/auth";
 import PropTypes from "prop-types";
 
 class App extends React.Component {
@@ -10,20 +10,6 @@ class App extends React.Component {
   static propTypes = {
     isLoggedIn: PropTypes.bool
   }
-
-  state = {
-    page: "home",
-    auth: false,
-  };
-
-  onLogin = (e) => {
-    e.preventDefault();
-    this.setState({ auth: true });
-  };
-
-  logout = () => {
-    this.setState({ auth: false });
-  };
 
   render() {
     return (
