+++ src/components/logo/LoftTaxi.jsx
@@ -2,11 +2,9 @@ import logo from "../../assets/img/company-logo.svg";
 
 const Logo = () => {
   return (
-    <>
-      <div className="company-logo">
-        <img src={logo} alt="lofttaxi"/>
-      </div>
-    </>
+    <div className="company-logo">
+      <img src={logo} alt="lofttaxi"/>
+    </div>
   );
 };
