+++ src/pages/Home.jsx
@@ -2,7 +2,7 @@ import { React, Component } from "react";
 import Map from "../components/map/Map";
 import Profile from "../components/profile/Profile";
 import Nav from "../components/nav/Nav";
-import { withAuth } from "../components/context/withAuth";
+import { withAuth } from "../context/auth";
 import PropTypes from "prop-types";
 
 class Home extends Component {
